+++ resources/views/existencias/cuentasporcobrar/index.blade.php
@@ -57,6 +57,8 @@
                                 <td>{{ $cc->pendiente }}</td>
                                 <td>{{ $cc->fecha }}</td>
                                 <td>
+                              @if(Auth::user()->rol!="Recepcionista")
+
                                     {!! Form::open(array(
                                         'style' => 'display: inline-block;',
                                         'method' => 'DELETE',
@@ -64,6 +66,7 @@
                                         'route' => ['admin.cuentasporcobrar.destroy', $cc->id])) !!}
                                     {!! Form::submit(trans('global.app_cob'), array('class' => 'btn btn-xs btn-danger')) !!}
                                     {!! Form::close() !!}
+                            @endif
                                 </td>
 
                             </tr>
